Feat(FE) : 선물 리스트 조회 등록 후 바로 적용되도록 수정
Former-commit-id: c464fd9afdf46f8bdcd505c3e3ca075183785b03

+++ frontend/src/components/present/PresentListSearch.vue
@@ -1,6 +1,6 @@
 <template>
 <div class="present-carousel">
-  <template v-if="cardList.length===0">
+  <template v-if="presentList.length===0">
     <v-img
     src="@/assets/img/none_present.png"
     >
@@ -97,6 +97,21 @@ export default {
   mounted(){
     this.search();
   },
+  watch:{
+    ...mapState(presentStore,["presentList"]),
+    presentList: {
+    deep: true,
+    handler(newVal) {
+      if(this.cardList.length === newVal.Present.length){
+        return;
+      }else if(this.cardList.length < newVal.Present.length){
+        this.cardList.push(newVal.Present[this.cardList.length]);
+      }else{
+        return;
+      } 
+    }
+  }
+  },
   computed:{
     ...mapState(presentStore,["presentList"]),
     columns(){
@@ -119,7 +134,34 @@ export default {
       this.dialog=false;
     },
     remove(index){
-      this.deletePresent(this.cardList[index].presentId);
+      this.$swal.fire({
+        title: '정말 삭제하시나요?',
+        text: "다시 한번 생각해주세요!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: '네, 삭제합니다!'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          let rmv=this.deletePresent(this.cardList[index].presentId);
+          if(rmv){
+            this.$swal.fire(
+              '삭제 성공!',
+              '파일이 삭제되었습니다.',
+              'success'
+            )
+            this.$delete(this.cardList,index);
+          }else{
+            this.$swal.fire(
+              '삭제 실패!',
+              '파일이 정상적으로 삭제되지 않았습니다.',
+              'error'
+            )
+          }
+        }
+      })
+      //
     }
   }
 }
